extract passport configuration
